Remove unsused files and refactor code
