+++ OTTCreator/OTTCreator.Client/MauiProgram.cs
@@ -1,6 +1,5 @@
 ﻿using CommunityToolkit.Maui;
 using Microsoft.Extensions.Logging;
-using OTTCreator.Client.Pages;
 
 namespace OTTCreator.Client
 {
@@ -18,9 +17,6 @@ namespace OTTCreator.Client
                     fonts.AddFont("OpenSans-Regular.ttf", "OpenSansRegular");
                     fonts.AddFont("OpenSans-Semibold.ttf", "OpenSansSemibold");
                 });
-            builder.Services.AddSingleton<ContentItemPage>();
-            builder.Services.AddSingleton<CategoryPage>();
-            builder.Services.AddSingleton<SupportPage>();
 #if DEBUG
             builder.Logging.AddDebug();
 #endif
