--- OTTCreator/OTTCreator.WebApp/Controllers/AccountController.cs
@@ -1,65 +1,0 @@
-﻿using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Mvc;
-using OTTCreator.WebApp.Areas.Identity.Data;
-using OTTCreator.WebApp.ViewModels;
-
-namespace OTTCreator.WebApp.Controllers;
-
-[Authorize]
-public class AccountController : Controller
-{
-    private readonly SignInManager<User> signInManager;
-
-    public AccountController(SignInManager<User> signInManager)
-    {
-        this.signInManager = signInManager;
-    }
-
-    [HttpGet]
-    [AllowAnonymous]
-    public IActionResult Login()
-    {
-        return View();
-    }
-
-    [HttpPost]
-    [AllowAnonymous]
-    [ValidateAntiForgeryToken]
-    public async Task<IActionResult> Login(LoginViewModel model)
-    {
-        var result = await signInManager.PasswordSignInAsync(model.UserName, model.Password, model.RememberMe, lockoutOnFailure: false);
-        if (result.Succeeded)
-        {
-            return RedirectToLocal("Home/Index");
-        }
-        else
-        {
-            ModelState.AddModelError(string.Empty, "Невдала спроба входу.");
-            return View(model);
-        }
-    }
-
-    [HttpPost]
-    [ValidateAntiForgeryToken]
-    public async Task<IActionResult> SignOff()
-    {
-        await signInManager.SignOutAsync();
-        return RedirectToAction("Login");
-    }
-
-    private IActionResult RedirectToLocal(string returnUrl)
-    {
-        if (Url.IsLocalUrl(returnUrl))
-            return Redirect(returnUrl);
-        else
-            return RedirectToAction(nameof(HomeController.Index), "Home");
-    }
-
-    [HttpGet]
-    public IActionResult AccessDenied(string returnUrl = null)
-    {
-        ViewData["ReturnUrl"] = returnUrl;
-        return View();
-    }
-}
