--- OTTCreator/OTTCreator.WebApp/Views/Account/AccessDenied.cshtml
@@ -1,12 +1,0 @@
-﻿@{
-    ViewData["Title"] = "Доступ заборонено";
-}
-
-<div class="card shadow mb-4">
-    <div class="card-header py-3">
-        <h6 class="m-0 font-weight-bold text-primary"> </h6>
-    </div>
-    <div class="card-body">
-        У вас недостатній рівень доступу для перегляду цієї сторінки.
-    </div>
-</div>
